Правки от наставника
